Get Club Stats

+++ composeApp/src/commonMain/kotlin/com/miuh/clubs/core/data/ClubsApi.kt
@@ -71,7 +71,7 @@ object ClubsApi {
 
 
     fun buildClubOverallStatsUrl(
-        genType: GenType, clubIds: List<Long>
+        genType: GenType, clubIds: List<Int>
     ): String {
         return URLBuilder().apply {
             takeFrom(BASE_URL)
